Split tests into multiple files for each module
